JBIDE-13995 EL suggestions box options depends on order of methods
The issue is fixed by the sorting of getters and setters and using an argument type as a property type in case
of setter is set as a first or the only one IJavaElement for the property.
The getters and setters are set to appear sorted in Content Assistatn additional info window as well as in EL Tooltips for
Java/HTML/JSP/XHTML Editors.
JUnit Test is added for the issue: org.jboss.tools.jsf.jsp.ca.test.CAELBeanPropertyTest.
Existing org.jboss.tools.jsf.jsp.hover.ELTooltipTest JUnit Test is updated as well due to test the issue.

	modified:   common/plugins/org.jboss.tools.common.el.core/src/org/jboss/tools/common/el/core/ca/AbstractELCompletionEngine.java
	modified:   common/plugins/org.jboss.tools.common.el.core/src/org/jboss/tools/common/el/core/resolver/TypeInfoCollector.java
	modified:   common/plugins/org.jboss.tools.common.el.ui/src/org/jboss/tools/common/el/ui/ca/ELProposalProcessor.java